Progress updates.

+++ examples/src/test/java/com/obsidiandynamics/indigo/ws/BaseHarness.java
@@ -3,6 +3,6 @@ package com.obsidiandynamics.indigo.ws;
 import java.util.concurrent.atomic.*;
 
 public abstract class BaseHarness implements AutoCloseable {
-  public final AtomicInteger received = new AtomicInteger();
-  public final AtomicInteger sent = new AtomicInteger();
+  public final AtomicLong received = new AtomicLong();
+  public final AtomicLong sent = new AtomicLong();
 }
